+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/Init.kt
@@ -7,14 +7,17 @@ import kotlinx.coroutines.launch
 import net.fabricmc.api.ModInitializer
 import net.minecraft.commands.arguments.EntityArgument
 import net.minecraft.commands.arguments.selector.EntitySelector
+import net.minecraft.network.chat.Component
 import tech.thatgravyboat.skyblockapi.api.SkyBlockAPI
 import tech.thatgravyboat.skyblockapi.api.events.base.Subscription
 import tech.thatgravyboat.skyblockapi.api.events.misc.RegisterCommandsEvent
 import tech.thatgravyboat.skyblockapi.helpers.McClient
+import tech.thatgravyboat.skyblockapi.helpers.McPlayer
 import tech.thatgravyboat.skyblockapi.utils.Logger
 import tech.thatgravyboat.skyblockpv.api.MojangAPI
 import tech.thatgravyboat.skyblockpv.command.SkyblockPlayerSuggestionProvider
 import tech.thatgravyboat.skyblockpv.screens.CollectionScreen
+import tech.thatgravyboat.skyblockpv.screens.MobScreen
 
 object Init : ModInitializer {
     override fun onInitialize() {
@@ -30,7 +33,7 @@ object Init : ModInitializer {
         event.register("pv") {
             callback {
                 McClient.tell {
-                    McClient.setScreen(CollectionScreen)
+                    McClient.setScreen(MobScreen(McPlayer.uuid))
                 }
             }
             then("player", StringArgumentType.string(), SkyblockPlayerSuggestionProvider) {
@@ -38,9 +41,12 @@ object Init : ModInitializer {
                     val player = this.getArgument("player", String::class.java)
                     CoroutineScope(Dispatchers.IO).launch {
                         val uuid = MojangAPI.getUUID(player)
-                        McClient.tell {
-                            // open the actual uuid screen here :3
-                            McClient.setScreen(CollectionScreen)
+                        if (uuid == null) {
+                            McPlayer.self?.displayClientMessage(Component.literal("§cPlayer could not be found"), false)
+                        } else {
+                            McClient.tell {
+                                McClient.setScreen(MobScreen(uuid))
+                            }
                         }
                     }
                 }
